[refactor] 오염 값을 다시 UInt 타입으로 바꿈 	이유: 출력할때 소수값이 포함된게 아닌 정수값이 필요해서

+++ iOS/DustNoticeApp/DustNoticeApp/Models/DustInfo.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct DustInfo: Codable {
-    let dustValue: Double
+    let dustValue: UInt
     let dustGrade: UInt
     let measureDate: Date
